+++ lib/config/DefaultSettings.ts
@@ -3,5 +3,6 @@ import type { PluginSettings } from './PluginSettings';
 export const DEFAULT_SETTINGS: PluginSettings = {
 	apiToken: null,
 	workspace: { id: 'none', name: 'None selected' },
-	charLimitStatusBar: 40
+	charLimitStatusBar: 40,
+	updateInRealTime: true
 };