+++ lib/constants.ts
@@ -2,4 +2,8 @@
  * The interval in ms at which the active timer is polled
  */
 export const ACTIVE_TIMER_POLLING_INTERVAL = 6000;
+
+/**
+ * The language string used for report code blocks.
+ */
 export const CODEBLOCK_LANG = 'toggl';